Core & Internals: re-organise transfer code
The motivation for this change comes from a circular import
dependency when I tried to import transfer.py into rules.
When a rule is cancelled, it needs to cancel the Transfer in
FTS. And it seems right to put the code which cancels the
transfer in transfer.py.

This commit only does code move: some of the functions which
don't need interaction with transfertools are moved to
request.py. RSE info retrieval is also moved out. What is
clearly only needed for submitters was moved to conveyor/
common.py.
