+++ lib/rucio/daemons/conveyor/stager.py
@@ -43,8 +43,7 @@ from rucio.common import exception
 from rucio.common.config import config_get, config_get_bool, config_get_int
 from rucio.common.logging import setup_logging
 from rucio.core.monitor import record_counter, record_timer
-from rucio.core import transfer as transfer_core
-from rucio.daemons.conveyor.common import submit_transfer, get_conveyor_rses, run_conveyor_daemon
+from rucio.daemons.conveyor.common import submit_transfer, get_conveyor_rses, run_conveyor_daemon, next_transfers_to_submit
 from rucio.db.sqla.constants import RequestType
 from rucio.transfertool.fts3 import FTS3Transfertool
 
@@ -53,7 +52,7 @@ graceful_stop = threading.Event()
 
 def run_once(bulk, group_bulk, rse_ids, scheme, failover_scheme, transfertool_kwargs, total_workers, worker_number, logger, activity):
     start_time = time.time()
-    transfers = transfer_core.next_transfers_to_submit(
+    transfers = next_transfers_to_submit(
         total_workers=total_workers,
         worker_number=worker_number,
         failover_schemes=failover_scheme,
